+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Environment Label Changelog
 
-_A plugin for Craft CMS 4.x to help distinguish your Craft environments ...so you don't forget where you are._
-
 The format of this file is based on ["Keep a Changelog"](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/). Version numbers follow the pattern: `MAJOR.FEATURE.BUGFIX`
+
+## 5.0.0 - 2024-05-23
+
+### Improved
+
+- Environment Label is ready for Craft 5!
+
+### Removed
+
+- Removed `getJs()`. The plugin no longer registers any info in JavaScript.
+
 
 ## 4.0.3 - 2023-04-25